Trait rfc::core::convert::cast::CastFrom

pub trait CastFrom<Src: ?Sized, Neglect = ()> where
    Neglect: CastOptions, {
    unsafe fn unsafe_cast_from(src: Src) -> Self
    where
        Src: Sized,
        Self: Sized,
        Neglect: CastOptions;

    fn cast_from(src: Src) -> Self
    where
        Src: Sized,
        Self: Sized,
        Neglect: SafeCastOptions,
    { ... }
}

🔬 This is a nightly-only experimental API. (cast)

Instantiate Self from a value of type Src.

The reciprocal of CastInto.

Required methods

unsafe fn unsafe_cast_from(src: Src) -> Self where
Src: Sized,
Self: Sized,
Neglect: CastOptions,

🔬 This is a nightly-only experimental API. (cast)

Instantiate Self by casting a value of type Src, potentially safely.

Loading content...
Provided methods

fn cast_from(src: Src) -> Self where
Src: Sized,
Self: Sized,
Neglect: SafeCastOptions,

🔬 This is a nightly-only experimental API. (cast)

Instantiate Self by casting a value of type Src, safely.

Loading content...