Trait rfc::core::convert::transmute::TransmuteFrom

pub unsafe trait TransmuteFrom<Src: ?Sized, Neglect = ()> where
    Neglect: TransmuteOptions, {
    fn transmute_from(src: Src) -> Self
    where
        Src: Sized,
        Self: Sized,
        Neglect: SafeTransmuteOptions,
    { ... }

    unsafe fn unsafe_transmute_from(src: Src) -> Self
    where
        Src: Sized,
        Self: Sized,
        Neglect: TransmuteOptions,
    { ... }
}

Reinterpret the bits of Src as a type Self.

The reciprocal of TransmuteFrom.

This trait is implemented automatically by the compiler for combinations of types where a transmutation is valid.

Provided methods

fn transmute_from(src: Src) -> Self where
Src: Sized,
Self: Sized,
Neglect: SafeTransmuteOptions,

Reinterpret the bits of a value of one type as another type, safely.

unsafe fn unsafe_transmute_from(src: Src) -> Self where
Src: Sized,
Self: Sized,
Neglect: TransmuteOptions,

Reinterpret the bits of a value of one type as another type, potentially unsafely.

The onus is on you to ensure that calling this function is safe.

Loading content...